full inference success

+++ examples/stable-diffusion/test_qwenimage_edit_vae.py
@@ -43,13 +43,13 @@ def test_vae(
     vae.eval()
 
     print("[2/4] Running PyTorch golden forward...")
-    dummy_enc_input = torch.randn(batch, 3, 1, height, width)
+    dummy_enc_input = torch.randn(batch, 3, 1, height, width, dtype=torch.float32)
     with torch.no_grad():
         golden_enc_output = vae._encode(dummy_enc_input)
 
     lat_h = height // vae_scale_factor
     lat_w = width // vae_scale_factor
-    dummy_dec_input = torch.randn(batch, z_dim, 1, lat_h, lat_w)
+    dummy_dec_input = torch.randn(batch, z_dim, 1, lat_h, lat_w, dtype=torch.float32)
     with torch.no_grad():
         golden_dec_output = vae._decode(dummy_dec_input, return_dict=False)
     if isinstance(golden_dec_output, tuple):
